Modified 'Save' button to show save form with placeholder text in OutputDisplay; add save and cancel buttons to save form; rename PatientData.js to OutputDisplay to better describe functionality
