+++ react-app/src/components/ResultPage.js
@@ -9,6 +9,7 @@ import '../stylesheets/ResultPage.scss';
 function ResultPage(props) {
   const [patientID, setPatientID] = useState(null);
   const [showLogs, setShowLogs] = useState(false);
+  const [showSaveForm, setShowSaveForm] = useState(false);
 
   return (
     <div>
@@ -20,14 +21,18 @@ function ResultPage(props) {
               loggedMessages={props.loggedMessages}
               setPatientID={setPatientID}
               setShowLogs={setShowLogs}
+              setShowSaveForm={setShowSaveForm}
             />
           </Col>
           <Col>
             <PatientData
               patientJson={props.extractedData[patientID]}
+              extractedData={props.extractedData}
               id={patientID}
               showLogs={showLogs}
               loggedMessages={props.loggedMessages}
+              showSaveForm={showSaveForm}
+              setShowSaveForm={setShowSaveForm}
             />
           </Col>
         </Row>
